amitai/requirements.txt
pygame
opencv-python
pywavefront
PyOpenGL